Remove as much warnings as possible
Project goes rom 200+ warnings to 7 due to tools and third parties, 0 warning on
project sources.

- remove unused parameters warnings by commenting the parameter name
- remove reorder warning by matching init order with declaration order
- remove make "jobserver unavailable" warning by using $(MAKE) i.e. make in .pro file
- fixed actual issues (e.g. missing return statement)
- also uses XPCF MACROs for attributes so that we can disable them at once and remove GCC
warnings on ignored atributes.

+++ src/base/geom/A2DPointsRectification.cpp
@@ -38,7 +38,7 @@ FrameworkReturnCode A2DPointsRectification::rectify(const std::vector<SolAR::dat
 	FrameworkReturnCode retCode = rectify(pts, camParams, rectParams, rectifiedPts);
 	if (retCode == FrameworkReturnCode::_SUCCESS) {
 		rectifiedKeypoints = keypoints;
-		for (int i = 0; i < rectifiedKeypoints.size(); ++i) {
+        for (size_t i = 0; i < rectifiedKeypoints.size(); ++i) {
 			rectifiedKeypoints[i].setX(rectifiedPts[i].getX());
 			rectifiedKeypoints[i].setY(rectifiedPts[i].getY());
 		}
